Handle regex strings with flags properly
E.g. /test/i is a valid regex that is accepted by Yara, but was previously not being handled properly. I've now added an extra member to YaraString which stores the regex flags, and adjusted the build method to properly build rules. The YaraBuilder now also accepts regex strings with flags.